requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.27.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
decorator==5.1.1
Deprecated==1.2.14
distro==1.8.0
fastapi==0.103.1
flake8==6.0.0
future==0.18.3
geocoder==1.38.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==6.0.0
isort==5.12.0
loguru==0.6.0
mccabe==0.7.0
mypy-extensions==1.0.0
netaddr==0.8.0
packaging==23.0
pathspec==0.11.0
Pillow==10.0.1
platformdirs==3.0.0
prometheus-client==0.16.0
psycopg2==2.9.5
pycodestyle==2.10.0
pydantic==2.3.0
pydantic_core==2.6.3
pyflakes==3.0.1
python-multipart==0.0.6
ratelim==0.1.6
requests==2.31.0
rfc3986==1.5.0
sentry-sdk==1.15.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
starlette==0.27.0
tabulate==0.9.0
tomli==2.0.1
typing_extensions==4.7.1
urllib3==1.26.14
uvicorn==0.20.0
win32-setctime==1.1.0
wrapt==1.15.0
yoyo-migrations==8.2.0
zipp==3.13.0
